Added a user input to the flask cli drop_tables command

+++ app/commands.py
@@ -7,11 +7,19 @@ from . import db
 cli = AppGroup('cli')
 
 
-@cli.command(name="create_tables")
+@cli.command(name='create_tables')
 def create_tables():
     db.create_all()
 
 
-@cli.command(name="drop_tables")
+@cli.command(name='drop_tables')
 def drop_tables():
-    db.drop_all()
+    print('You are about to drop all tables from your database')
+    response = input('Are you sure? (y/n): ')
+    if response == 'Y' or 'y' or 'Yes' or 'yes':
+        print('Dropping all tables...')
+        db.drop_all()
+        return
+
+    print('Did not drop all tables')
+    return
